refactor(pivot): silnik do bpp/pivot/, multiseek_registry re-eksportuje
Silnik tabeli krzyżowej przestał być multiseek-specyficzny w momencie, gdy
dostała ją też strona „Wyszukiwanie zapytaniem", a zadania 8-10 dokładają
drugi rejestr wymiarów (dla modelu Autor, z trzema bazami agregacji). Kod
rozdzielony na generyczny silnik (bpp/pivot/core.py) i rejestr wymiarów
Rekordu (bpp/pivot/rekord.py); multiseek_registry/pivot.py zostaje cienkim
re-eksportem, żeby nie ruszać call-site'ów w szablonach i widokach.

Jedyna zmiana sygnatury: zbuduj_pivot(..., *, model=None) — strategia dedupu
potrzebuje modelu bazowego, dotąd zaszytego jako Rekord. Domyślny None
zachowuje wszystkie istniejące wywołania; zadanie 8 poda tu Autora.

Progi PIVOT_MAX_CELLS/PIVOT_MAX_PAIRS świadomie NIE są re-eksportowane przez
shim: zbuduj_pivot czyta je jako globalne swojego modułu, więc kopia nazwy
pod starą ścieżką byłaby martwa — monkeypatch na niej przestawiłby kopię, nie
bramkę. Trzy testy patchujące progi celują teraz w bpp.pivot.core, gdzie te
stałe mieszkają. Brak nazwy w shimie jest tu cechą, nie luką: monkeypatch
domyślnie podnosi AttributeError na nieistniejącym atrybucie, więc pomyłka
wyjdzie głośno w linii patcha, zamiast dać fałszywie zielony test.

Neutralność: 25 testów pivota multiseeka przechodzi bez zmian w asercjach
(zmienione tylko trzy cele monkeypatcha — patch site idzie za przeprowadzką
stałej dokładnie tak jak import site). Cała suita pivot/zapytanie/multiseek:
458 passed.
